@@ -1,35 +1,42 @@
 {
-  "name": "sttv",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "cross-env NODE_ENV=development vite",
-    "build": "vite build",
-    "build:watch": "vite build --watch",
-    "test": "jest",
-    "test:watch": "jest --watchAll",
-    "serve": "vite preview"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.37",
-    "@testing-library/svelte": "^3.0.3",
-    "@types/jest": "^26.0.21",
-    "autoprefixer": "^10.2.5",
-    "babel-jest": "^26.6.3",
-    "capture-video-frame": "^1.0.0",
-    "cross-env": "^7.0.3",
-    "cssnano": "^5.0.6",
-    "downloadjs": "^1.4.7",
-    "jest": "^26.6.3",
-    "postcss": "^8.4.6",
-    "postcss-load-config": "^3.0.1",
-    "socket.io-client": "^4.4.1",
-    "svelte": "^3.35.0",
-    "svelte-canvas": "^0.7.1",
-    "svelte-jester": "^1.3.1",
-    "svelte-preprocess": "^4.6.9",
-    "tailwindcss": "^3.0.23",
-    "ts-jest": "^26.5.4",
-    "typescript": "^4.2.3",
-    "vite": "^2.8.4"
-  }
+	"name": "new-client",
+	"version": "0.0.1",
+	"scripts": {
+		"dev": "svelte-kit dev",
+		"build": "svelte-kit build",
+		"package": "svelte-kit package",
+		"preview": "svelte-kit preview",
+		"prepare": "svelte-kit sync",
+		"test": "playwright test",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
+		"format": "prettier --write --plugin-search-dir=. ."
+	},
+	"devDependencies": {
+		"@playwright/test": "^1.22.2",
+		"@sveltejs/adapter-auto": "next",
+		"@sveltejs/kit": "next",
+		"@typescript-eslint/eslint-plugin": "^5.27.0",
+		"@typescript-eslint/parser": "^5.27.0",
+		"autoprefixer": "^10.4.4",
+		"capture-video-frame": "^1.0.0",
+		"downloadjs": "^1.4.7",
+		"eslint": "^8.16.0",
+		"eslint-config-prettier": "^8.3.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"postcss": "^8.4.12",
+		"postcss-load-config": "^3.1.4",
+		"prettier": "^2.6.2",
+		"prettier-plugin-svelte": "^2.7.0",
+		"socket.io-client": "^4.5.1",
+		"svelte": "^3.44.0",
+		"svelte-canvas": "^0.7.1",
+		"svelte-check": "^2.7.1",
+		"svelte-preprocess": "^4.10.5",
+		"tailwindcss": "^3.0.23",
+		"tslib": "^2.3.1",
+		"typescript": "^4.7.2"
+	},
+	"type": "module"
 }
